Separated pandas integration
